+++ benchmarks/storage/bin/bench.dart
@@ -23,35 +23,54 @@ void main() async {
     ..init('.')
     ..registerAdapter(TestModelAdapter());
 
+  await ce.IsolatedHive.init('.');
+  await ce.IsolatedHive.registerAdapter(TestModelAdapter());
+
   await Isar.initialize('assets/libisar_macos.dylib');
 
   v4.Hive.defaultDirectory = '.';
   v4.Hive.registerAdapter('TestModel', (json) => TestModel.fromJson(json));
 
   final ceResults = <BenchResult>[];
+  final isolatedResults = <BenchResult>[];
   final v4Results = <BenchResult>[];
 
   for (final operations in benchmarks) {
     final ceResult = await runBenchmark(
+      name: 'Hive CE',
       operations: operations,
       type: DbType.hive,
       openBox: ce.Hive.openBox,
     );
 
+    final isolatedResult = await runBenchmark(
+      name: 'IsolatedHive',
+      operations: operations,
+      type: DbType.hive,
+      openBox: ce.IsolatedHive.openBox,
+    );
+
     final v4Result = await runBenchmark(
+      name: 'Hive v4',
       operations: operations,
       type: DbType.isar,
       openBox: (name) => v4.Hive.box(name: name, maxSizeMiB: 1024),
     );
 
     ceResults.add(ceResult);
+    isolatedResults.add(isolatedResult);
     v4Results.add(v4Result);
   }
+
+  await ce.Hive.close();
+  await ce.IsolatedHive.close();
+  v4.Hive.closeAllBoxes();
 
   final csv = const ListToCsvConverter().convert([
     [
       'Operations',
       'Hive CE Time',
+      'IsolatedHive Time',
       'Hive CE Size',
       'Hive v4 Time',
       'Hive v4 Size',
@@ -60,6 +79,7 @@ void main() async {
       [
         benchmarks[i],
         formatTime(ceResults[i].time),
+        formatTime(isolatedResults[i].time),
         formatSize(ceResults[i].size),
         formatTime(v4Results[i].time),
         formatSize(v4Results[i].size),