+++ hive/test/util/is_browser/is_browser.dart
@@ -1,0 +1,2 @@
+export 'impl/is_browser_vm.dart'
+    if (dart.library.js_interop) 'impl/is_browser_js.dart';