+++ hive_generator/example/lib/hive/hive_registrar.g.dart
@@ -23,3 +23,21 @@ extension HiveRegistrar on HiveInterface {
     registerAdapter(NullableTypesAdapter());
   }
 }
+
+extension IsolatedHiveRegistrar on IsolatedHiveInterface {
+  Future<void> registerAdapters() {
+    return Future.wait([
+      registerAdapter(Class1Adapter()),
+      registerAdapter(Class2Adapter()),
+      registerAdapter(ClassSpec1Adapter()),
+      registerAdapter(ClassSpec2Adapter()),
+      registerAdapter(ConstructorDefaultsAdapter()),
+      registerAdapter(EmptyClassAdapter()),
+      registerAdapter(Enum1Adapter()),
+      registerAdapter(EnumSpecAdapter()),
+      registerAdapter(IterableClassAdapter()),
+      registerAdapter(NamedImportsAdapter()),
+      registerAdapter(NullableTypesAdapter()),
+    ]);
+  }
+}