+++ hive_generator/lib/src/builder/registrar_builder.dart
@@ -90,6 +90,20 @@ extension HiveRegistrar on HiveInterface {
     buffer.write('''
   }
 }
+
+extension IsolatedHiveRegistrar on IsolatedHiveInterface {
+  Future<void> registerAdapters() {
+    return Future.wait([
+''');
+
+    for (final adapter in adapters) {
+      buffer.writeln('      registerAdapter($adapter()),');
+    }
+
+    buffer.write('''
+    ]);
+  }
+}
 ''');
 
     var registrarLocation = 'lib';