+++ hive/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## NEXT
 
+- Isolate support through `IsolatedHive`
+- Warning messages for potentially unsafe isolate usage
 - Updates suggested placement of `GenerateAdapters` annotation
 
 ## 2.10.1